refactor(uniffi): extract fire-uniffi-notifications
Phase 3b of the uniffi multi-namespace split.

Rust:
- Add rust/crates/fire-uniffi-notifications with
  setup_scaffolding!("fire_uniffi_notifications"). Hosts FireNotificationsHandle
  plus the notification record types. All 15 methods previously inlined on
  FireAppCore (notifications, bookmarks, drafts, read-history,
  set_topic_notification_level) move here verbatim.
- FireAppCore gains notifications: Arc<FireNotificationsHandle> field plus
  notifications() getter. state_notification.rs is deleted.
- Promote cross-domain types to fire-uniffi-types so the notifications crate
  can reference them without depending on fire-uniffi (which would create a
  cycle): TopicListState + its nested TopicRowState, TopicSummaryState,
  TopicUserState, TopicPosterState, TopicParticipantState, TopicTagState
  (needed by fetch_bookmarks / fetch_read_history), and the draft trio
  DraftState, DraftDataState, DraftListResponseState (needed by the drafts
  APIs). Records are organised into fire-uniffi-types/src/records/{tag,
  topic_list, draft}.rs. state_topic_list.rs and state_topic_detail.rs now
  keep only the types that remain topic-domain-specific.

iOS / Android:
- FireSessionStore on both platforms routes notification/bookmark/draft/
  read-history calls through core.notifications(). Other call sites go
  through FireSessionStore wrappers and need no change.
- Kotlin imports: NotificationCenterState / NotificationListState move to
  uniffi.fire_uniffi_notifications; TopicListState / TopicRowState /
  TopicSummaryState / TopicTagState / TopicPosterState / TopicUserState
  move to uniffi.fire_uniffi_types. Updated FireSessionStore.kt,
  MainActivity.kt, TopicPresentation.kt, and TopicPresentationTest.kt
  accordingly.

Build plumbing:
- native/ios-app/project.yml preBuildScript outputs gain
  fire_uniffi_notifications.swift. Regenerated project.pbxproj.

cargo test --workspace stays green.
