+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "rust/crates/fire-core",
     "rust/crates/fire-uniffi",
     "rust/crates/fire-uniffi-diagnostics",
+    "rust/crates/fire-uniffi-notifications",
     "rust/crates/fire-uniffi-search",
     "rust/crates/fire-uniffi-types",
 ]
