+++ rust/crates/fire-uniffi-types/src/records/mod.rs
@@ -1,0 +1,10 @@
+pub mod draft;
+pub mod tag;
+pub mod topic_list;
+
+pub use draft::{DraftDataState, DraftListResponseState, DraftState};
+pub use tag::RequiredTagGroupState;
+pub use topic_list::{
+    TopicListState, TopicParticipantState, TopicPosterState, TopicRowState, TopicSummaryState,
+    TopicTagState, TopicUserState,
+};
