+++ rust/crates/fire-uniffi-types/src/records/tag.rs
